changed path for print logo

+++ htmlstaticcontent/changePathToimg.py
@@ -4,30 +4,44 @@ import sys,os,re
 
 _PATH = os.path.abspath(os.path.dirname(__file__))
 
-filename = "028_products_krovl_vodostochnye_sistemy_/index.html"
+dirList = os.listdir(_PATH)
+# print dirList
 
-html_file_name = os.path.join(_PATH, filename)
-html_file = open(html_file_name,'r')
-html_file_string = html_file.read()
-html_file.close()
+for dirname in dirList: 
+ 	dirpath = os.path.join(_PATH, dirname)
+	filename = os.path.join(dirpath, 'index.html')
+	# print filename
+	if os.path.exists(filename):
+		# print filename
+		html_file_name = filename
+		html_file = open(html_file_name,'r')
+		html_file_string = html_file.read()
+		html_file.close()
 
-p6 = re.compile("/upload[\w\-\s_/\.]+jpg|/upload[\w\-\s_/\.]+png|/upload[\w\-\s_/\.]+gif|/upload[\w\-\s_/\.]+ico")
+		stringToReplace = "/bitrix/templates/trimet/img/default/p5.png"
+		stringForReplace = "/img/p5.png"
 
-imgpaths = p6.findall(html_file_string)
+		html_file_string = html_file_string.replace(stringToReplace, stringForReplace)
 
-for path in imgpaths:
-    p4 = re.compile("[\w\-_\s\.]+jpg|[\w\-_\s\.]+png|[\w\-_\s\.]+gif|[\w\-_\s\.]+ico")
-    imgname = p4.findall(path)
+		print "done on ", filename
 
-    for img in imgname:
+		# p6 = re.compile("/upload[\w\-\s_/\.]+jpg|/upload[\w\-\s_/\.]+png|/upload[\w\-\s_/\.]+gif|/upload[\w\-\s_/\.]+ico")
 
-        html_file_string = html_file_string.replace(path, "/htmlstaticcontent/unknown_images/"+img)
+		# imgpaths = p6.findall(html_file_string)
 
-        print path, " --> ", img
+		# for path in imgpaths:
+		#     p4 = re.compile("[\w\-_\s\.]+jpg|[\w\-_\s\.]+png|[\w\-_\s\.]+gif|[\w\-_\s\.]+ico")
+		#     imgname = p4.findall(path)
+
+		#     for img in imgname:
+
+		#         html_file_string = html_file_string.replace(path, "/htmlstaticcontent/unknown_images/"+img)
+
+		#         print path, " --> ", img
 
 
-html_file_name = os.path.join(_PATH, filename)
-html_file = open(html_file_name,'w')
+		html_file_name = filename
+		html_file = open(html_file_name,'w')
 
-html_file.write(html_file_string)
-html_file.close()
+		html_file.write(html_file_string)
+		html_file.close()
